Refactored code, make objects immutable.
